Adding 1a.c changes

+++ Codeforces/1a.c
@@ -1,8 +1,8 @@
 #include<stdio.h>
 int main()
 {
-int n,m,a,ntile=0,mtile=0;
-scanf("%d %d %d",&n,&m,&a);
+long long int n,m,a,ntile=0,mtile=0;
+scanf("%lld %lld %lld",&n,&m,&a);
 if(n%a!=0)
 ntile=(n/a) +1;
 else
@@ -11,5 +11,5 @@ if(m%a!=0){
 mtile=(m/a)*ntile;}
 else{
 mtile=((m/a)-1)*ntile;}
-printf("%d\n",ntile+mtile);
+printf("%lld\n",ntile+mtile);
 }
